generationLLM/app/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile
#
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   starlette
    #   watchfiles
audioop-lts==0.2.1
    # via
    #   speechrecognition
    #   standard-aifc
certifi==2025.4.26
    # via requests
cffi==1.17.1
    # via cryptography
charset-normalizer==3.4.2
    # via
    #   pdfminer-six
    #   requests
click==8.2.1
    # via uvicorn
cryptography==45.0.3
    # via pdfminer-six
dnspython==2.7.0
    # via pymongo
fastapi==0.115.12
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
fpdf==1.7.2
    # via -r requirements.in
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
gridfs==0.0.1
    # via -r requirements.in
h11==0.16.0
    # via uvicorn
hf-xet==1.1.3
    # via huggingface-hub
httptools==0.6.4
    # via uvicorn
huggingface-hub==0.32.4
    # via
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   requests
jinja2==3.1.6
    # via torch
markupsafe==3.0.2
    # via jinja2
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via torch
numpy==2.2.6
    # via transformers
packaging==25.0
    # via
    #   huggingface-hub
    #   pytesseract
    #   transformers
pdfminer-six==20250327
    # via pdfplumber
pdfplumber==0.11.6
    # via -r requirements.in
pillow==11.2.1
    # via
    #   -r requirements.in
    #   pdfplumber
    #   pytesseract
pycparser==2.22
    # via cffi
pydantic==2.11.5
    # via fastapi
pydantic-core==2.33.2
    # via pydantic
pymongo==4.13.0
    # via -r requirements.in
pypdfium2==4.30.1
    # via pdfplumber
pytesseract==0.3.13
    # via -r requirements.in
python-dotenv==1.1.0
    # via uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   huggingface-hub
    #   transformers
    #   uvicorn
regex==2024.11.6
    # via transformers
requests==2.32.3
    # via
    #   huggingface-hub
    #   transformers
safetensors==0.5.3
    # via transformers
sniffio==1.3.1
    # via anyio
speechrecognition==3.14.3
    # via -r requirements.in
standard-aifc==3.13.0
    # via speechrecognition
standard-chunk==3.13.0
    # via standard-aifc
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
tokenizers==0.21.1
    # via transformers
torch==2.7.1
    # via -r requirements.in
tqdm==4.67.1
    # via
    #   huggingface-hub
    #   transformers
transformers==4.52.4
    # via -r requirements.in
typing-extensions==4.14.0
    # via
    #   fastapi
    #   huggingface-hub
    #   pydantic
    #   pydantic-core
    #   speechrecognition
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.4.0
    # via requests
uvicorn[standard]==0.34.3
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.0.5
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
